This is a template for creating OpenLCB projects for the ARM Cortex M3 architecture. Currently configured only for LPC1769, should work fine with LPC1768 as well. Additional chips should be easy to add.
This is just a first stab. The OpenLCB and CAN libraries are currently untested, but should work. The demo only makes the LPCXpresso LED blink, which is enough for now, as it proves the damned thing works.

To use, have a look at the makefile. Editing it should be easy enough. You can also load the project into Eclipse (currently only tested with Eclipse Helios + Code Sourcery GCC Lite on OS X 10.6), and compile.

The makefiles will generate two versions, one suitable for loading directly onto an LPCXpresso or other board via your favorite ISP, JTAG device, or the LPC-LINK. The other is to be used with the NXP USB secondary bootloader (as modified by CodeRed, and modified a little bit more by me. I will be posting the code to GitHub soon, after I finish polishing it up; in the meantime if you want the code or a hex file to burn just ask and I will provide).

Mad props to Per Eklund, and to this dude: http://gvworks.blogspot.com/

More, much more, to come! This code will be merged with the Arduino code---somehow---to make a unified version suitable for multiple architectures.