Added BG interface.
Press Blue to start teaching.
or
Press Gold to start learning.

In either mode, use Blue to iterate through the OUTPUTS. The output LED will come on to indicate that it is selected; Blue will be on solid to indicate that "on" will be taught/learned or will flash to indicate that "off" will be taught/learned.
In either mode, press an input button to teach/learn that input "on"; press a second time to teac/learn "off" instead; press a third time to cancel teaching/learning that input.
In either mode, press Gold to finalize, or press both simultanesouly to cancel everything.

Press and hold B and G until the pretty lights come on, and then back off to do a factory reset (currently, only reboots, doesn't wipe the stored EventIDs yet, for testing purposes).